bot_requirements.txt
opencv-python==4.9.0.80
osmnx==1.8.1
rasterio==1.3.9
python_dotenv==1.0.0
aiogram==2.25.1